requirements.txt

evidently
fastapi
google-cloud-storage
joblib
jupyter
matplotlib
mlflow
numpy
pandas
prometheus-fastapi-instrumentator==5.7.1
psycopg2-binary 
pyarrow
pydantic
python-dotenv
requests
scikit-learn
sqlalchemy
tqdm
uvicorn


